requirements.txt
fastapi
uvicorn
pydantic
pyjwt
python-jose
python-dotenv
python-multipart
jwt
python-jwt